+++ src/main/java/com/pawntracker/config/WebSecurityConfig.java
@@ -34,8 +34,7 @@ public class WebSecurityConfig  extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
       http
                     .authorizeRequests()
-              .antMatchers("/admin/**").hasRole("ADMIN")
-                .antMatchers("/owner/**").hasRole("OWNER")
+              .antMatchers("/admin/**").hasAnyRole("ADMIN", "OWNER")
               .antMatchers(
                       "/",
                       "/favicon.ico",
