+++ src/main/resources/templates/admin/layout/sidebar_menu.html
@@ -9,19 +9,19 @@
             </li>
 
             <li class="nav-item">
-                <a class="nav-link" href="/owner/">
+                <a class="nav-link" href="/admin/">
                     <span data-feather="file"></span>
                     Main
                 </a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/owner/items/all">
+                <a class="nav-link" href="/admin/items/all">
                     <span data-feather="file"></span>
                     All Items
                 </a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/owner/items/toapprove">
+                <a class="nav-link" href="/admin/items/toapprove">
                     <span data-feather="shopping-cart"></span>
                     Items to approve
                 </a>
